Move rotation code into its own file for specs that didn't already do that
